Removed build-in linters
