--- test-fixtures/eslint/.eslintrc.json
@@ -1,21 +1,0 @@
-{
-  "env": {
-    "es6": true,
-    "node": true
-  },
-  "plugins": ["prettier"],
-  "rules": {
-    "prettier/prettier": [
-      "error",
-      {
-        "singleQuote": true,
-        "trailingComma": "es5",
-        "semi": false
-      }
-    ],
-    "indent": ["error", 4],
-    "linebreak-style": ["error", "unix"],
-    "quotes": ["error", "single"],
-    "semi": ["error", "never"]
-  }
-}
