--- test-fixtures/eslint/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "prettier.eslintIntegration": true,
-    "prettier.singleQuote": false
-}
