--- test-fixtures/eslint/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "testeslint",
-  "version": "1.0.0",
-  "description": "Test eslint integration on this dummy project",
-  "private": true,
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "cigit",
-  "license": "MIT",
-  "devDependencies": {
-    "eslint": "^5.16.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "prettier": "^1.18.2"
-  },
-  "dependencies": {}
-}
