--- test-fixtures/eslint/withEslint.js
@@ -1,4 +1,0 @@
-// Settings (eslint): single-quote, trailing-comma, no-semi
-function foo(){
-    return "bar";
-}
