--- test-fixtures/tslint/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "prettier.tslintIntegration": true,
-    "prettier.singleQuote": false
-}
