--- test-fixtures/tslint/withTslint.ts
@@ -1,4 +1,0 @@
-// Settings (tslint): single-quote, trailing-comma, no-semi
-function foo(){
-    return "bar";
-}
